requirements.txt

fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
qdrant-client==1.12.1
openai==1.51.2
httpx==0.27.2
pydantic==2.9.2
pdfplumber==0.11.4
python-docx==1.1.2
python-multipart==0.0.12
sqlite-utils==3.37
pytest==8.3.3
